Refactor in a similar pattern to existing official nu plugins
