+++ src/lib.rs
@@ -1,0 +1,4 @@
+mod nu;
+mod weather;
+
+pub use weather::Weather;
